updated VocalVideo component and Testimonials.js imports, VocalVideo component now embedded in Testimonials component. Also updated package.json with installed dependencies, react-slick and slick-carousel
