+++ src/components/Testimonials.js
@@ -1,15 +1,14 @@
 import React from "react";
 import './Testimonials.css';
-import Bertie from '../assets/images/Bertrand-Icon.png';
-import VideoEmbed from "./VideoEmbed";
+import VocalVideo from "./VocalVideo.js";
 function Testimonials() {
-  console.log(Bertie);
+ 
   return (
     <section className="jumbotron testimonials" id="testimonials">
       <div className="container">
         <h1 className="display-4 text-center test-video section-heading">Testimonials</h1>
 
-    <VideoEmbed/>
+    <VocalVideo/>
 
       </div>
     </section>
